4.21 complete

+++ part4/blog-list/controllers/blogs.js
@@ -41,8 +41,18 @@ blogsRouter.post('/', async (request, response) => {
 blogsRouter.delete('/:id', async (request, response) => {
     const id = request.params.id
 
-    await Blog.findByIdAndDelete(id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, config.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const blog = await Blog.findById(id)
+    if (blog.user.toString() === decodedToken.id) {
+        await Blog.findByIdAndDelete(id)
+        return response.status(204).end()
+    } else {
+        return response.status(400).json({ error: 'only the creator can delete the blog' })
+    }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
